backend/DOLMA/requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
google-api-core==1.26.3
google-auth==1.28.0
google-cloud-speech==2.2.0
google-cloud-texttospeech==2.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
libcst==0.3.18
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.0.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
